psyq-obj-parser: Handle comm symbols properly
Convert .comm symbols into the proper ELF equivalent, instead of
converting them to local symbol in the .bss segment.

This makes it possible use the converted PSY-Q libraries to recreate a
bit-exact executable of Frogger - without proper comm symbol support
controlling the order of some .bss symbols in the libraries is
impossible.

Even though this is more correct than before there is a chance some
users of psyq-obj-parser are relying on the old behaviour thus a command
line option is added which can be used to restore the previous
behaviour.